==> jupyter_notebooks/01_SSM_1km_V1.ipynb <==
import sys
!conda install -c conda-forge --yes --prefix {sys.prefix} xarray, rasterio, tqdm, scikit-learn, geemap, rioxarray

import sys
!conda install --yes --prefix {sys.prefix} scikit-image

import sys
!pip install --prefix {sys.prefix} pikepdf, calplot, openeo

# Import modules.
from irrigation_detection import search_files
from irrigation_detection import subset_ds
from irrigation_detection import export_values
import xarray as xr

# Find all .nc files nested inside directories first argument is path where files and directories stored second argument is the file ending.
files = search_files('D://ssm_download/',".nc") 

# Create a xarray object from multiple .nc files.
ssm = xr.open_mfdataset(files, decode_cf=True)

ssm

import hvplot.xarray

test = ssm.ssm.sel(lon='8.4067', lat='49.643', method='nearest')

test.hvplot.scatter()

# Subset and mask out unrelevant areas.
ssm_subset = subset_ds(r'C:\Users\USER\Desktop\Master_Irrigation\03_GIS\gg_test_field\gg_test_field_1.shp', ssm)

# Apply conversation of digital units to physical units.
ssm_subset['ssm'] = ssm_subset['ssm'] * 0.5
ssm_subset['ssm_noise'] = ssm_subset['ssm_noise'] * 0.5

import 

# Export xr as .nc file 
ssm_subset[['ssm', 'ssm_noise']].to_netcdf(r'C:\Users\USER\Desktop\Master_Irrigation\03_GIS\datasets\SSM_1km_V1.nc')

export_values(latitude=49.85083645182146, longitude=8.463941352976518, ds=ssm, name='point_2_ssm.csv')
