==> jupyter_notebooks/02_RADOLAN_1km.ipynb <==
Radolan Specs
NODATA_value -1
Unit 1/10 mm

Radolan right after load:
    min: 0 max: 4095
    nan occurence:  False
Radolan after Subsetting and convert digital to physical values:
    min:0.0 max: 123.5
    nan occurence:  True

from irrigation_detection import download_radolan_SF
from irrigation_detection import unpack_radolan_SF
from irrigation_detection import open_radolan_SF
from irrigation_detection import subset_ds
from irrigation_detection import export_values

# Download Radolan SF Products from https://opendata.dwd.de/. Arguments must be (startyear, stopyear, local download folder).
download_radolan_SF(2015, 2021, 'D://radolanSF_download/')

# Exctract files from archive.
unpack_radolan_SF('D://radolanSF_download')

# Read radolan files into an xarray dataset.
radolan_SF = open_radolan_SF('D://radolanSF_download/*2350-dwd---bin.gz')

# Subset dataset
radolan_SF_subset = subset_ds(r'C:\Users\USER\Desktop\Master_Irrigation\03_GIS\oberrheingraben_shp\oberrheingraben-polygon.shp', radolan_SF)

# Apply conversation of digital units to physical units.
radolan_SF_subset['SF'] = radolan_SF_subset['SF'] * 0.1

# Export xr as .nc file 
radolan_SF_subset.to_netcdf(r'C:\Users\USER\Desktop\Master_Irrigation\03_GIS\datasets\RADOLAN_1km.nc')

# ROI Data
from irrigation_detectionigation_detectionigation_detectionrigation_detectionrigation_detectionrigation_detectionrigation_detection import 
radolan_SF_subset = subset_ds(r'C:\Users\USER\Desktop\Master_Irrigation\03_GIS\oberrheingraben_shp\oberrheingraben-polygon.shp', radolan_SF)

import xarray as xr
import hvplot.xarray
radolan = xr.open_dataset(r'C:\Users\USER\Desktop\Master_Irrigation\03_GIS\datasets\RADOLAN_1km.nc')

rain_per_year = radolan.SF.groupby("time.year").sum()
p1 = rain_per_year.hvplot.quadmesh(groupby='year',clim=(0,800),cmap='gist_rainbow', crs='WGS84',frame_height=540, projection='PlateCarree',project=True, geo=True, rasterize=True, dynamic=False)
p1

def xr_var_stats(da):
    """
    Arguments:
    """
    
    import xarray as xr
    import numpy as np

    #get unique values
    unique_values = np.unique(da)
    
    #Print min,max, nan occurence, 
    print('min: %s' %unique_values.min(), 'max: %s' %unique_values.max())
    if True in np.isnan(unique_values):
        print('nan occurence:  True')
    else:
        print('nan occurence:  False')
    
    return print('stats finished')

xr.plot.hist(radolan_SF_subset.SF.isel(time=1), bins=20, range=(0.1,40), edgecolor='black')

export_values(latitude=49.85083645182146, longitude=8.463941352976518, ds=radolan_SF_subset, name='point_1_SF.csv')

#Visualisations with http://christopherbull.com.au/mkmov/ command line tool

radolan_SF_subset = radolan_SF_subset.load()

radolan_SF_subset

hvplot.save(radolan_SF_subset.isel(time=slice(0,10)).hvplot.quadmesh(crs='WGS84',frame_height=540, projection='PlateCarree',project=True, geo=True, rasterize=True, dynamic=False), "test2.html")
