==> jupyter_notebooks/05_BDF5L.ipynb <==
import sys
!conda install -c conda-forge --yes --prefix {sys.prefix} owslib rtree 

import wfs server in qgis and save data as shapefile   
RLP: https://www.geoportal.rlp.de/linked_open_data/   
Hessen: https://www.hlnug.de/fileadmin/dokumente/boden/BFD5L/methoden/m46.html   

TOPSOIL Organic Carbon https://esdac.jrc.ec.europa.eu/tmp_dataset_access_req_50307#tabs-0-filters=1    


import geopandas as gpd

# Load shapefile exported by qgis and transform it from EPSG:25832 (ETRS89 / UTM zone 32N) to EPSG:4326 (WGS84).
bfd5l_m46 = gpd.read_file(r'C:\Users\USER\Desktop\Master_Irrigation\03_GIS\BFDL_ML46\BFD5L_ML46.shp').to_crs(epsg=4326)

# Load aoi for further clipping of soil data.
aoi = gpd.read_file(r'C:\Users\USER\Desktop\Master_Irrigation\03_GIS\oberrheingraben_shp\oberrheingraben-polygon.shp')

# Subset soil data.
bfd5l_m46_subset = bfd5l_m46.clip(aoi, keep_geom_type=True)

# Export Subset to disk
bfd5l_m46_subset.to_file(r'C:\Users\USER\Desktop\Master_Irrigation\03_GIS\BFDL_ML46\BFD5L_subset.shp')

# Read LUCAS TOPSOIL data for points in Hessisches Ried.
p_ried = gpd.read_file(r'C:\Users\USER\Desktop\Master_Irrigation\03_GIS\p_ried\LTS_p_ried.shp')

# Spatial join to find soilclass for LUCAS TOPSOIL points.
p_ried = p_ried.sjoin(bfd5l_m46_subset)

# Update shapefile from boden export
#p_ried.to_file(r'C:\Users\USER\Desktop\Master_Irrigation\03_GIS\p_ried\LTS_p_ried_BDF5L.shp')

import hvplot.pandas

# Lets see if points are in polygon.
test = bfd5l_m46_subset.loc[[276,209146,136604,205862]]

# Plot soil class polygons behind p_ried spatial joined gdf.
test.hvplot(geo=True, frame_width=600, tiles='EsriImagery',) * p_ried.hvplot(geo=True, color='orange')
