==> jupyter_notebooks/06_Evapo_rel_pot.ipynb <==
Die Werte im Raster müssen durch 10 dividiert werden, um die richtigen Werte in mm zu bekommen.    
array([-9999,     0,     1,     2,     3,     4,     5,     6,     7,
           8,     9,    10,    11,    12,    13,    14,    15,    16]) -> unkonvertiert    
https://epsg.io/31467    

from irrigation_detection import download_evapo
from irrigation_detection import extract_evapo
from irrigation_detection import subset_ds
from irrigation_detection import open_evapo

# Download real evaporation to disk
download_evapo(2015, 2021,'D://evapo_r_download', real=True)

# Download potential evaporation to disk
download_evapo(2015, 2021,'D://evapo_p_download', real=False)

# Use Gdal Warp to transform .asc with EPSG:31467 to .tif with EPSG:4326
extract_evapo('D://evapo_r_download/')

# Use Gdal Warp to transform .asc with EPSG:31467 to .tif with EPSG:4326
extract_evapo('D://evapo_p_download/')

# Read tiff files from disk.
evapo_r = open_evapo('D://evapo_r_download/data/', real=True)

evapo_r_subset = subset_ds(r'C:\Users\USER\Desktop\Master_Irrigation\03_GIS\oberrheingraben_shp\oberrheingraben-polygon.shp', evapo_r)

# fill invalid pixels (-999) with nan
evapo_r_subset = evapo_r_subset.where(evapo_r_subset.evapo_r >= 0, drop=True)
# Apply conversation of digital units to physical units.
evapo_r_subset['evapo_r'] = evapo_r_subset['evapo_r'] * 0.1

# Export netcdf file to disk
evapo_r_subset.to_netcdf('C://Users/USER/Desktop/Master_Irrigation/03_GIS/datasets/EVAPOTRANSPIRATION_REAL.nc')

# Read tiff files from disk.
evapo_p = open_evapo('D://evapo_p_download/data/', real=False)

evapo_p_subset = subset_ds(r'C:\Users\USER\Desktop\Master_Irrigation\03_GIS\oberrheingraben_shp\oberrheingraben-polygon.shp', evapo_p)

# fill invalid pixels (-999) with nan
evapo_p_subset = evapo_p_subset.where(evapo_p_subset.evapo_p >= 0, drop=True)
# Apply conversation of digital units to physical units.
evapo_p_subset['evapo_p'] = evapo_p_subset['evapo_p'] * 0.1

# Export netcdf file to disk
evapo_p_subset.to_netcdf('C://Users/USER/Desktop/Master_Irrigation/03_GIS/datasets/EVAPOTRANSPIRATION_POTENTIAL.nc')
