==> jupyter_notebooks/10_PYSMM_surface_soil_moisture.ipynb <==
import sys
!conda install -c conda-forge --yes --prefix {sys.prefix} geemap

import geopandas as gpd
from shapely.geometry import Point
from pysmm.derive_SM import get_ts

# Define AOI

path = 'C://Users/USER/Desktop/Master_Irrigation/03_GIS/idm_test/whr_test_field.geojson'
gdf = gpd.read_file(path)
gdf["x"] = gdf.centroid.x
gdf["y"] = gdf.centroid.y

# Use PYSMM to get data for AOI

from pysmm.derive_SM import get_map
get_map(gdf.bounds.minx.values[0], gdf.bounds.miny.values[0], gdf.bounds.maxx.values[0], gdf.bounds.maxy.values[0],
        'idm_gee_export',
        sampling=50,                             # specify desired output resolution
        year=None, month=None, day=None,        # define a date (optional)
        tracknr=None,                           # define the Sentinel-1 track (optinal)
        overwrite=False,
        start='2018-03-01',
        stop='2018-03-30') 

#To derive the soil moisture time-series of a single location use
sm_ts = get_ts((gdf.x.values[0], gdf.y.values[0]),
              'C://Users/USER/Desktop/Master_Irrigation/03_GIS/idm_test/get_ts/',
              footprint=50,
              calc_anomalies=False,
              create_plots=False,
              export_csv='whr_test_field_data')
